test runtime build

+++ pallets/ocex/rpc/runtime-api/src/lib.rs
@@ -34,8 +34,8 @@ sp_api::decl_runtime_apis! {
 		// Returns the asset inventory deviation in the offchain State
 		fn calculate_inventory_deviation() -> Result<BTreeMap<AssetId,Decimal>, sp_runtime::DispatchError>;
 		// Retrieve a sorted vector of accounts for a given epoch and market based on descending order of scores
-		fn top_lmp_accounts(epoch: u32, market: TradingPair, sorted_by_mm_score: bool, limit: u16) -> Vec<AccountId>;
+		fn top_lmp_accounts(epoch: u16, market: TradingPair, sorted_by_mm_score: bool, limit: u16) -> Vec<AccountId>;
 		// Returns the eligible rewards for given main, epoch and market
-		fn calculate_lmp_rewards(main: AccountId, epoch: u32, market: TradingPair) -> (Decimal, Decimal, bool);
+		fn calculate_lmp_rewards(main: AccountId, epoch: u16, market: TradingPair) -> (Decimal, Decimal, bool);
 	}
 }
